backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
python-multipart==0.0.6
alembic==1.12.1
python-jose==3.3.0
passlib==1.7.4
python-dateutil==2.8.2
bcrypt==4.0.1
httpx==0.25.2
